fix: install split apk if not installed

+++ root_util.sh
@@ -23,15 +23,17 @@ fi
 [ -d /data/adb/post-fs-data.d/ ] || mkdir -p /data/adb/post-fs-data.d/
 [ -d /data/adb/service.d/ ] || mkdir -p /data/adb/service.d/
 
-rm "/data/adb/post-fs-data.d/umount_revanced_$pkgName.sh"
-rm "/data/adb/service.d/mount_revanced_$pkgName.sh"
-rm "/data/local/tmp/revancify/$pkgName.apk"
+[ -e "/data/adb/post-fs-data.d/umount_revanced_$pkgName.sh" ] && rm "/data/adb/post-fs-data.d/umount_revanced_$pkgName.sh"
+[ -e "/data/adb/service.d/mount_revanced_$pkgName.sh" ] && rm "/data/adb/service.d/mount_revanced_$pkgName.sh"
+[ -e "/data/local/tmp/revancify/$pkgName.apk" ] && rm "/data/local/tmp/revancify/$pkgName.apk"
 
 
-if pm list packages | grep -q "$pkgName" && [ "$(dumpsys package "$pkgName" | sed -n '/versionName/s/.*=//p' | sed 's/ /./1p')" = "$appVer" ]; then
-    :
-else
-    pm install --user 0 -r "apps/$appName-$appVer.apk"
+if ! (pm list packages | grep -q "$pkgName" && [ "$(dumpsys package "$pkgName" | sed -n '/versionName/s/.*=//p' | sed 's/ /./1p')" = "$appVer" ]); then
+    if [ -d "apps/$appName-$appVer" ]; then
+        pm install --user 0 -r "apps/$appName-$appVer/"*
+    else
+        pm install --user 0 -r "apps/$appName-$appVer.apk"
+    fi
 fi
 
 pm list packages | grep -q "$pkgName" || exit 1
@@ -40,6 +42,7 @@ stockApp=$(pm path "$pkgName" | sed -n "/base/s/package://p")
 revancedApp="/data/local/tmp/revancify/$pkgName.apk"
 
 am force-stop "$pkgName"
+pm clear --cache-only "$pkgName"
 
 {
     grep "$pkgName" /proc/mounts | cut -d " " -f 2 | sed "s/apk.*/apk/" | xargs -r umount -vl
@@ -50,6 +53,7 @@ am force-stop "$pkgName"
 } > /storage/emulated/0/Revancify/install_log.txt 2>&1
 
 am force-stop "$pkgName"
+pm clear --cache-only "$pkgName"
 
 grep -q "$pkgName" /proc/mounts || exit 1
